requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx
python-dotenv
pydantic
